[BEAM-12727] extract Concatenate CombineFn to runner-core module to deduplicate it from many runners
